+++ java/client/src/main/java/glide/api/commands/servermodules/FT.java
@@ -162,7 +162,9 @@ public class FT {
      * @return A two element array, where first element is count of documents in result set, and the
      *     second element, which has format <code>
      *     {@literal Map<GlideString, Map<GlideString, GlideString>>}</code> - a mapping between
-     *     document names and map of their attributes.<br>
+     *     document names and map of their attributes. When {@link
+     *     FTSearchOptions.FTSearchOptionsBuilder#nocontent()} is set, the attribute maps will be
+     *     empty.<br>
      *     If {@link FTSearchOptions.FTSearchOptionsBuilder#count()} or {@link
      *     FTSearchOptions.FTSearchOptionsBuilder#limit(int, int)} with values <code>0, 0</code> is
      *     set, the command returns array with only one element - the count of the documents.